chains-202105 scripts

+++ web/chains-202105/bin/stop.sh
@@ -1,5 +1,10 @@
 #! /usr/bin/env bash
 set -o errexit -o errtrace -o pipefail -o nounset
 cd $(dirname $0)/..
-pkill -F gunicorn.pid
-sleep 3
+if [[ -f  gunicorn.pid ]]; then
+    pkill -F gunicorn.pid
+    sleep 3
+else
+    echo No pidfile gunicorn.pid >&2
+    exit 1
+fi
